+++ docs/changelog.rst
@@ -17,6 +17,7 @@ New features
 * Class :class:`bibrecord.record.Dataset` for datasets.
 * Method :meth:`bibrecord.record.Record.from_bib` to read/populate records from BibTeX entries.
 * Module :mod:`bibrecord.bibtex` for basic handling of BibTeX bibliography files.
+* Module :mod:`bibrecord.database` for handling multiple bibliographic records.
 
 
 Version 0.1.0